% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotics_rescaling.R
\name{abiotics_rescaling}
\alias{abiotics_rescaling}
\title{Title Rescale abiotics data to 10 or 30 arcmins}
\usage{
abiotics_rescaling(
  flo1k_data,
  worldclim_data,
  earthenv_data,
  minlat,
  maxlat,
  minlong,
  maxlong,
  resolution = 30,
  geosphere = FALSE
)
}
\arguments{
\item{flo1k_data}{raster stack with data from FLO1K}

\item{worldclim_data}{raster stack with data from WorldClim}

\item{earthenv_data}{raster stack with data from EarthEnv}

\item{minlat}{Float; Mininum latitude}

\item{maxlat}{Float; Maximum latitude}

\item{minlong}{Float; Minimum longitude}

\item{maxlong}{Float; Maximum longitude}

\item{resolution}{Integer; Chosen resolution (10 or 30, default 30)}

\item{geosphere}{Boolean; If TRUE, import data from geosphere (default=FALSE, takes a long time)}
}
\value{
Returns a rescaled abiotics dataframe
}
\description{
Title Rescale abiotics data to 10 or 30 arcmins
}
