% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAndDownload.R
\name{checkAndDownload}
\alias{checkAndDownload}
\title{Title Download datasets from various databases if the files do not already exist on hard drive}
\usage{
checkAndDownload(url, filePath)
}
\arguments{
\item{url}{String; Url to download}

\item{filePath}{String; Output path of the downloaded file}
}
\description{
Title Download datasets from various databases if the files do not already exist on hard drive
}
