% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{data_cleaning}
\alias{data_cleaning}
\title{Title Remove suspicious occurence points}
\usage{
data_cleaning(data, minlat, maxlat, minlong, maxlong, check.out = TRUE)
}
\arguments{
\item{data}{A data frame downloaded from GBIF}

\item{minlat}{Minimum latitude}

\item{maxlat}{Maximum latitude}

\item{minlong}{Minimum longitude}

\item{maxlong}{Maximum longitude}

\item{check.out}{Check outliers or no (take a long time for a large number of occurences)}
}
\value{

}
\description{
Title Remove suspicious occurence points
}
