% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_monoculture_app.R
\name{design_monoculture_app}
\alias{design_monoculture_app}
\title{Title Plot species density diagrams for individual species}
\usage{
design_monoculture_app(
  species_df,
  abiotics_df,
  species_list,
  minlat = -56,
  maxlat = 60,
  minlong = -145,
  maxlong = 180
)
}
\arguments{
\item{species_list}{}
}
\value{

}
\description{
Title Plot species density diagrams for individual species
}
