% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_DB.R
\name{download_DB}
\alias{download_DB}
\title{Title Download datasets from various databases}
\usage{
download_DB(databases_to_use, resolution)
}
\arguments{
\item{databases_to_use}{List of databases to use}

\item{resolution}{Resolution}
}
\value{

}
\description{
Title Download datasets from various databases
}
