% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarthEnv_select.R
\name{EarthEnv_select}
\alias{EarthEnv_select}
\title{Title Extract chosen paramters (annual_meanT, T_seasonality, maxT_WM, minT_CM, annual_rangeT, meanT_DQ, ph_max, elevation_avg, slope_avg) from EarthEnv}
\usage{
EarthEnv_select()
}
\value{
A raster brick containing EarthEnv data
}
\description{
Title Extract chosen paramters (annual_meanT, T_seasonality, maxT_WM, minT_CM, annual_rangeT, meanT_DQ, ph_max, elevation_avg, slope_avg) from EarthEnv
}
