% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalfspherical.R
\name{evalfspherical}
\alias{evalfspherical}
\title{Title}
\usage{
evalfspherical(
  data,
  radius,
  points,
  getid.nearestneighbor = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{radius}{Float; radius to search}

\item{getid.nearestneighbor}{Boolean; affected to nothing ? (default=FALSE)}

\item{verbose}{Boolean (default=TRUE)}
}
\description{
Title
}
