% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLO1K_select.R
\name{FLO1K_select}
\alias{FLO1K_select}
\title{Title Exctract chosen paramters from FLO1K}
\usage{
FLO1K_select(resolution)
}
\arguments{
\item{resolution}{Resolution}
}
\value{

}
\description{
Title Exctract chosen paramters from FLO1K
}
