% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_species_hypervolumes.R
\name{generate_species_hypervolumes}
\alias{generate_species_hypervolumes}
\title{Title}
\usage{
generate_species_hypervolumes(
  species_list_selected,
  rescaled_abiotics,
  nb_axis
)
}
\arguments{
\item{species_list_selected}{List of the selected species}

\item{rescaled_abiotics}{Selected and rescaled abiotics pca axis in a data frame}

\item{nb_axis}{Number of axis retained in the pca}
}
\value{

}
\description{
Title
}
