% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HV_intersection.R
\name{HV_intersection}
\alias{HV_intersection}
\title{Title Calculate the volume of the intersection of the hypervolumes}
\usage{
HV_intersection(
  hv_list,
  num.points.max = NULL,
  verbose = TRUE,
  distance.factor = 1
)
}
\arguments{
\item{hv_list}{List of hypervolumes}

\item{num.points.max}{Integer; The maximum number of points to keep for each hypervolume (default=NULL, the number of point is calculated by ceiling(10^(3 + sqrt(hv_list[\link{i}]@Dimensionality))))}

\item{verbose}{Boolean; Print the calculated maximum number of points and the minimal density (default=TRUE)}

\item{distance.factor}{Integer; Used to calculate if two points on two volumes are overlapping : cutoff_dist = point_density^(-1/dim) * distance.factor (default=1)}
}
\value{
Returns a Hypervolume containing @Method, @Data, @Dimensionality, @Volume, @PointDensity, @Parameters, @RandomPoints and @ValueAtRandomPoints
}
\description{
Title Calculate the volume of the intersection of the hypervolumes
}
