% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypervolume_set_n_intersection.R
\name{hypervolume_set_n_intersection}
\alias{hypervolume_set_n_intersection}
\title{Title Calculate the volume of the intersection of the hypervolumes}
\usage{
hypervolume_set_n_intersection(
  hv_list,
  num.points.max = NULL,
  verbose = FALSE,
  distance.factor = 1,
  check.hyperplane = FALSE
)
}
\arguments{
\item{hv_list}{List of hypervolumes}

\item{num.points.max}{Integer; The maximum number of points to keep for each hypervolume (default=NULL, the number of point is calculated by ceiling(10^(3 + sqrt(hv_list[\link{i}]@Dimensionality))))}

\item{verbose}{Boolean; Returns the calculated maximum number of points, the minimal density and the progression (default=FALSE)}

\item{distance.factor}{Integer; Used to calculate if two points on two volumes are overlapping : cutoff_dist = point_density^(-1/dim) * distance.factor (default=1)}

\item{check.hyperplane}{Boolean; If true, check if the data is hyperplanar (default=FALSE)}
}
\value{
Returns a Hypervolume containing @Method, @Data, @Dimensionality, @Volume, @PointDensity, @Parameters, @RandomPoints and @ValueAtRandomPoints
}
\description{
Title Calculate the volume of the intersection of the hypervolumes
}
