% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_user_system.R
\name{input_user_system}
\alias{input_user_system}
\title{Title Register the parameters corresponding to user's system}
\usage{
input_user_system(abiotics_df)
}
\arguments{
\item{abiotics_df}{data frame with abiotics data}
}
\value{

}
\description{
Title Register the parameters corresponding to user's system
}
