% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kdtree_ball_query_id_multiple}
\alias{kdtree_ball_query_id_multiple}
\title{Title}
\usage{
kdtree_ball_query_id_multiple(tr, ptlist, nr, nc, r, verb)
}
\arguments{
\item{verb}{}
}
\value{

}
\description{
Title
}
