% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kdtree_range_query_multiple}
\alias{kdtree_range_query_multiple}
\title{Title}
\usage{
kdtree_range_query_multiple(tr, pminlist, pmaxlist, nr, nc, verb)
}
\arguments{
\item{verb}{}
}
\value{

}
\description{
Title
}
