% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occ.R
\name{plot_occ}
\alias{plot_occ}
\title{Title Plot retained and remove occurence points per species}
\usage{
plot_occ(data, data_cl)
}
\arguments{
\item{data_cl}{}
}
\value{

}
\description{
Title Plot retained and remove occurence points per species
}
