% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pca_axis.R
\name{set_pca_axis}
\alias{set_pca_axis}
\title{Title Convert dataframe into PCA projection}
\usage{
set_pca_axis(species_abiotics_df)
}
\arguments{
\item{species_abiotics_df}{data frame with merged species and abiotics}
}
\value{

}
\description{
Title Convert dataframe into PCA projection
}
