% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_system.R
\name{set_system}
\alias{set_system}
\title{Title Project user's system into PCA space}
\usage{
set_system(species_abiotics_df, user_param_syst)
}
\arguments{
\item{user_param_syst}{}
}
\value{

}
\description{
Title Project user's system into PCA space
}
