% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldClim_select.R
\name{WorldClim_select}
\alias{WorldClim_select}
\title{Title Extract chosen parameters (read srad, vapr and prec) from WorldClim}
\usage{
WorldClim_select()
}
\value{
Returns a raster brick containing WorldClim data
}
\description{
Title Extract chosen parameters (read srad, vapr and prec) from WorldClim
}
