% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotics_rescaling.R
\name{abiotics_rescaling}
\alias{abiotics_rescaling}
\title{Title Rescale abiotics data to 10 or 30 arcmins}
\usage{
abiotics_rescaling(
  flo1k_data,
  worldclim_data,
  earthenv_data,
  minlat,
  maxlat,
  minlong,
  maxlong,
  resolution,
  geosphere = FALSE
)
}
\arguments{
\item{flo1k_data}{raster stack with data from FLO1K}

\item{worldclim_data}{raster stack with data from WorldClim}

\item{earthenv_data}{raster stack with data from EarthEnv}

\item{minlat}{Mininum latitude}

\item{maxlat}{Maximum latitude}

\item{minlong}{Minimum longitude}

\item{maxlong}{Maximum longitude}

\item{resolution}{Resolution}

\item{geosphere}{If TRUE, import data from geosphere, which take a lot of time}
}
\value{

}
\description{
Title Rescale abiotics data to 10 or 30 arcmins
}
\examples{

}
