% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_density.R
\name{design_monoculture_app}
\alias{design_monoculture_app}
\title{Title Plot species density diagrams for individual species}
\usage{
design_monoculture_app(abiotics_df, species_list)
}
\arguments{
\item{species_list}{}
}
\value{

}
\description{
Title Plot species density diagrams for individual species
}
