% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_monoculture_app.R
\name{design_monoculture_app}
\alias{design_monoculture_app}
\title{Title Plot species density diagrams for individual species}
\usage{
design_monoculture_app(
  species_df,
  abiotics_df,
  species_list,
  minlat = -56,
  maxlat = 60,
  minlong = -145,
  maxlong = 180
)
}
\arguments{
\item{minlat}{Float; Minimum latitude (default=-56)}

\item{maxlat}{Float; Maximum latitude (default=60)}

\item{minlong}{Float; Minimum longitude (default=-145)}

\item{maxlong}{Float; Maximum longitude (default=180)}
}
\description{
Title Plot species density diagrams for individual species
}
