% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_polyculture_app.R
\name{design_polyculture_app}
\alias{design_polyculture_app}
\title{Title App to design polyculture system}
\usage{
design_polyculture_app(rescaled_combi_df, species_abiotics_df)
}
\arguments{
\item{rescaled_combi_df}{Data frame with rescaled compatibility index}

\item{species_abiotics_df}{dataframe with all abiotic variables}
}
\value{

}
\description{
Title App to design polyculture system
}
