% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarthEnv_select.R
\name{EarthEnv_select}
\alias{EarthEnv_select}
\title{Title Exctract chosen paramters from EarthEnv}
\usage{
EarthEnv_select()
}
\value{

}
\description{
Title Exctract chosen paramters from EarthEnv
}
\examples{

}
