% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalfspherical.R
\name{evalfspherical}
\alias{evalfspherical}
\title{Title}
\usage{
evalfspherical(
  data,
  radius,
  points,
  getid.nearestneighbor = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
