% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLO1K_select.R
\name{FLO1K_select}
\alias{FLO1K_select}
\title{Title Extract chosen paramters (all) from FLO1K}
\usage{
FLO1K_select(resolution = 30)
}
\arguments{
\item{resolution}{Integer; Chosen resolution (10 or 30, default 30)}
}
\value{
A raster brick containing FLO1K data
}
\description{
Title Extract chosen paramters (all) from FLO1K
}
