% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_download.R
\name{gbif_download}
\alias{gbif_download}
\title{downloading gbif data}
\usage{
gbif_download(species_list)
}
\arguments{
\item{species_list}{The list of species to be considered (Scientific names)}

\item{gbif_user}{Username on GBIF}

\item{gbif_pwd}{Password on GBIF}

\item{mail}{email address used for GBIF}
}
\value{
A data frame with all the occurrence points of the species selected
}
\description{
downloading gbif data
}
