% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_species_hypervolumes.R
\name{generate_species_hv}
\alias{generate_species_hv}
\title{Title Generate a list of hypervolume from a list of species}
\usage{
generate_species_hv(species_list, rescaled_abiotics)
}
\arguments{
\item{rescaled_abiotics}{Selected and rescaled abiotics pca axis in a data frame}

\item{species_list_selected}{List of the selected species}

\item{nb_axis}{Number of axis retained in the pca}
}
\value{

}
\description{
Title Generate a list of hypervolume from a list of species
}
