% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HV_intersection.R
\name{HV_intersection}
\alias{HV_intersection}
\title{Title}
\usage{
HV_intersection(
  hv_list,
  num.points.max = NULL,
  verbose = TRUE,
  distance.factor = 1
)
}
\arguments{
\item{hv_list}{List of hypervolumes}

\item{num.points.max}{Integer}

\item{verbose}{Boolean}

\item{distance.factor}{Integer}
}
\value{

}
\description{
Title
}
