% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_rescaling.R
\name{occ_rescaling}
\alias{occ_rescaling}
\title{Title}
\usage{
occ_rescaling(data_cl, minlat, maxlat, minlong, maxlong)
}
\arguments{
\item{data_cl}{GBIF cleaned data frame}

\item{minlat}{Minimum latitude}

\item{maxlat}{Maximum latitude}

\item{minlong}{Minimum longitude}

\item{maxlong}{Maximum longitude}

\item{resolution}{Resolution}
}
\value{

}
\description{
Title
}
