% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_rescaling.R
\name{occ_rescaling}
\alias{occ_rescaling}
\title{Title Rescale occurence data at 10 or 30 arcmins}
\usage{
occ_rescaling(data_cl, resolution)
}
\arguments{
\item{data_cl}{GBIF cleaned data frame}

\item{resolution}{Resolution}
}
\value{

}
\description{
Title Rescale occurence data at 10 or 30 arcmins
}
