% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_rescaling.R
\name{occ_rescaling}
\alias{occ_rescaling}
\title{Title Rescale species coordinates and/or abiotic coordinates at 10 or 30 arcmins}
\usage{
occ_rescaling(data_cl, resolution = 30)
}
\arguments{
\item{data_cl}{species coordinates and/or abiotic coordinates dataframe}

\item{resolution}{Integer; Chosen resolution (10 or 30, default 30)}
}
\value{
Returns a rescaled coordinates dataframe
}
\description{
Title Rescale species coordinates and/or abiotic coordinates at 10 or 30 arcmins
}
