% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pca_axis.R
\name{set_pca_axis}
\alias{set_pca_axis}
\title{Title}
\usage{
set_pca_axis(species_abiotics_df, nb_variables)
}
\arguments{
\item{species_abiotics_df}{data frame with species and abiotics}

\item{nb_variables}{number of selected variables}
}
\value{

}
\description{
Title
}
