% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_compatibility_app.R
\name{system_compatibility_app}
\alias{system_compatibility_app}
\title{Title App to test if the user's aquaculture system is suitable to a species or combination}
\usage{
system_compatibility_app(
  hv_list,
  abiotics_df,
  user_param_syst,
  user_param_rescaled,
  species_abiotics_df,
  species_list
)
}
\arguments{
\item{hv_list}{a list of hypervolumes}

\item{abiotics_df}{data frame with abiotics data}

\item{user_param_syst}{user parameters}

\item{user_param_rescaled}{user parameters rescaled}

\item{species_abiotics_df}{Dataframe; Species abiotic data}

\item{species_list}{List; List of species name}
}
\value{
Returns nothing
}
\description{
Title App to test if the user's aquaculture system is suitable to a species or combination
}
