% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldClim_select.R
\name{WorldClim_select}
\alias{WorldClim_select}
\title{Title Exctract chosen paramters from WorldClim}
\usage{
WorldClim_select()
}
\value{

}
\description{
Title Exctract chosen paramters from WorldClim
}
